+++ controllers/MotionMergingTrait.php
@@ -314,6 +314,7 @@ trait MotionMergingTrait
                 $newMotion,
                 array_map('intval', $this->getHttpRequest()->post('amendStatus', [])),
                 $this->getHttpRequest()->post('newStatus'),
+                $this->getHttpRequest()->post('newSubstatus'),
                 $this->getHttpRequest()->post('newInitiator', ''),
                 $this->getHttpRequest()->post('votes', []),
                 $this->getHttpRequest()->post('amendVotes', [])