+++ History.md
@@ -12,6 +12,7 @@
 - When choosing "Tags / categories" as home page layout, then the consolidated category-list at the top can now be deactivated.
 - It can now be set up so that the modified version proposed as part of a proposed procedure is shown inline as part of the motion.
 - Besides of exporting all motions as a ZIP-file containing single ODT files, a single ODT file containing all motion texts can now be exported.
+- When a new version of a motion is created during merging amendments (in contrast to a resolution being created), the status of the new motion version can be explicitly specified.
 - The ODT export now also supports numbered lists.
 - The motion list can now be filtered for To Do items (that is, motions/amendments that need to be screened) and also shows the To Do action for items on that list as part of the Status.
 - The motion list now persists its filter and sort settings for each user session, until changed or reset.